@@ -88,6 +88,9 @@ public class ZoomView  implements PlatformView,
             case "logout":
                 logout(result);
                 break;
+            case "get_meeting_password":
+                getMeetingPassword(result);
+                break;
             default:
                 result.notImplemented();
         }
@@ -201,6 +204,13 @@ public class ZoomView  implements PlatformView,
             }
         }
     }
+    
+    private void getMeetingPassword(MethodChannel.Result result){
+        InMeetingService mInMeetingService = ZoomSDK.getInstance().getInMeetingService();
+        System.out.println(mInMeetingService.getCurrentMeetingNumber());
+        System.out.println(mInMeetingService.getMeetingPassword());
+        result.success(Arrays.asList(String.valueOf(mInMeetingService.getCurrentMeetingNumber()),mInMeetingService.getMeetingPassword()));
+    }
 
     private void startMeeting(MethodCall methodCall, MethodChannel.Result result) {
 
@@ -238,10 +248,57 @@ public class ZoomView  implements PlatformView,
         result.success(true);
     }
 
-    private boolean parseBoolean(Map<String, String> options, String property, boolean defaultValue) {
-        return options.get(property) == null ? defaultValue : Boolean.parseBoolean(options.get(property));
-    }
+    private void startInstantMeeting(MethodCall methodCall, MethodChannel.Result result) {
+        Map<String, String> options = methodCall.arguments();
+        ZoomSDK zoomSDK = ZoomSDK.getInstance();
 
+        if(!zoomSDK.isInitialized()) {
+            System.out.println("Not initialized!!!!!!");
+            result.success(998);
+            return;
+        }
+
+        // if(!zoomSDK.isLoggedIn()) {
+        //     result.success(999);
+        //     return;
+        // }
+
+        final MeetingService meetingService = zoomSDK.getMeetingService();
+
+        StartMeetingOptions opts = new StartMeetingOptions();
+        opts.custom_meeting_id = options.get("customMeetingId");
+        opts.invite_options = Integer.parseInt(options.get("inviteOptions")); 
+        opts.meeting_views_options = Integer.parseInt(options.get("meetingViewsOptions")); 
+        opts.no_bottom_toolbar = parseBoolean(options, "noBottomToolbar", false); 
+        opts.no_chat_msg_toast = parseBoolean(options, "noChatMsgToast", false); 
+        opts.no_dial_in_via_phone = parseBoolean(options, "noDialInViaPhone", false); 
+        opts.no_dial_out_to_phone = parseBoolean(options, "noDialOutToPhone", false); 
+        opts.no_disconnect_audio = parseBoolean(options, "noDisconnectAudio", false); 
+        opts.no_driving_mode = parseBoolean(options, "noDrivingMode", false); 
+        opts.no_invite = parseBoolean(options, "noInvite", false); 
+        opts.no_meeting_end_message = parseBoolean(options, "noMeetingEndMessage", false); 
+        opts.no_meeting_error_message = parseBoolean(options, "noMeetingErrorMessage", false); 
+        opts.no_share = parseBoolean(options, "noShare", false); 
+        opts.no_titlebar = parseBoolean(options, "noTitlebar", false); 
+        opts.no_unmute_confirm_dialog = parseBoolean(options, "noUnmuteConfirmDialog", false); 
+        opts.no_video = parseBoolean(options, "noVideo", false); 
+        opts.no_webinar_register_dialog = parseBoolean(options, "noWebinarRegisterDialog", false); 
+        opts.participant_id = options.get("participantId"); 
+        opts.no_audio = parseBoolean(options, "noAudio", false); 
+
+            int ret = meetingService.startInstantMeeting(context, opts);
+
+        System.out.println("SFLINK SDK FLUTTER: startInstantMeeting ret: " + ret);
+
+        InMeetingService mInMeetingService = zoomSDK.getInMeetingService();
+        System.out.println("SFLINK SDK FLUTTER: get meetingID" );
+
+        System.out.println(mInMeetingService.getCurrentMeetingNumber());
+
+        System.out.println(mInMeetingService.getMeetingPassword());
+        result.success(ret);
+    }
+   
 
     private void meetingStatus(MethodChannel.Result result) {
 
@@ -265,19 +322,38 @@ public class ZoomView  implements PlatformView,
     }
 
     private void loginWithEmail(MethodCall methodCall, MethodChannel.Result result) {
-        result.success(Arrays.asList("ZOOM_LOGIN_SUCCESS", "No status available"));
+        Map<String, String> options = methodCall.arguments();
+        ZoomSDK zoomSDK = ZoomSDK.getInstance();
+        if(zoomSDK.isLoggedIn()) {
+            result.success(0);
+        }else{
+            result.success(zoomSDK.loginWithZoom(options.get("email"),options.get("password")));
+        }
+       
     }
     
     private void loginWithSso(MethodCall methodCall, MethodChannel.Result result) {
-        result.success(Arrays.asList("ZOOM_LOGIN_SUCCESS", "No status available"));
+        Map<String, String> options = methodCall.arguments();
+        ZoomSDK zoomSDK = ZoomSDK.getInstance();
+        if(zoomSDK.isLoggedIn()) {
+            System.out.println("ZOOM SDK FLUTTER : isLoggedIn ");
+            result.success(zoomSDK.tryAutoLoginZoom());
+        }else{
+            result.success(zoomSDK.loginWithSSOToken(options.get("token")));
+        }
     }
     
-    private void startInstantMeeting(MethodCall methodCall, MethodChannel.Result result) {
-        result.success(Arrays.asList("ZOOM_START_MEETING_SUCCESS", "Meeting Id", "Meeting pass"));
-    }
-   
+    
     private void logout(MethodChannel.Result result) {
+        ZoomSDK zoomSDK = ZoomSDK.getInstance();
+        if(zoomSDK.isLoggedIn()) {
+            zoomSDK.logoutZoom();
+        }
         result.success(true);
+    }
+
+    private boolean parseBoolean(Map<String, String> options, String property, boolean defaultValue) {
+        return options.get(property) == null ? defaultValue : Boolean.parseBoolean(options.get(property));
     }
 
     @Override
@@ -285,7 +361,7 @@ public class ZoomView  implements PlatformView,
 
     @Override
     public void onZoomAuthIdentityExpired() {
-
+        System.out.println("ZOOM SDK FLUTTER : expired 1");
     }
 
     @Override
