@@ -1,7 +1,7 @@
-import 'package:flutter_zoom_plugin_example/meeting_screen.dart';
+import './meeting_screen.dart';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_zoom_plugin_example/start_meeting_screen.dart';
+import './start_meeting_screen.dart';
 
 class JoinWidget extends StatefulWidget {
   @override
