@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import 'package:flutter_zoom_plugin_example/join_screen.dart';
-import 'package:flutter_zoom_plugin_example/meeting_screen.dart';
-import 'package:flutter_zoom_plugin_example/start_meeting_screen.dart';
+import './join_screen.dart';
+// import 'package:flutter_zoom_plugin_example/meeting_screen.dart';
+// import 'package:flutter_zoom_plugin_example/start_meeting_screen.dart';
 
 void main() => runApp(ExampleApp());
 
@@ -18,8 +18,8 @@ class ExampleApp extends StatelessWidget {
       initialRoute: '/',
       routes: {
         '/': (context) => JoinWidget(),
-        '/meeting': (context) => MeetingWidget(),
-        '/startmeeting': (context) => StartMeetingWidget(),
+        // '/meeting': (context) => MeetingWidget(),
+        // '/startmeeting': (context) => StartMeetingWidget(),
       },
     );
   }
